+++ backend/apid/routers/silenced.go
@@ -25,11 +25,11 @@ func NewSilencedRouter(store store.Store) *SilencedRouter {
 // Mount the SilencedRouter to a parent Router
 func (r *SilencedRouter) Mount(parent *mux.Router) {
 	routes := resourceRoute{router: parent, pathPrefix: "/silenced"}
-	routes.index(r.list)
-	routes.show(r.find)
-	routes.create(r.create)
-	routes.update(r.update)
-	routes.destroy(r.destroy)
+	routes.getAll(r.list)
+	routes.get(r.find)
+	routes.post(r.create)
+	routes.del(r.destroy)
+	routes.put(r.createOrReplace)
 
 	// Custom
 	routes.path("subscriptions/{subscription}", r.list).Methods(http.MethodGet)
@@ -60,13 +60,13 @@ func (r *SilencedRouter) create(req *http.Request) (interface{}, error) {
 	return cfg, err
 }
 
-func (r *SilencedRouter) update(req *http.Request) (interface{}, error) {
+func (r *SilencedRouter) createOrReplace(req *http.Request) (interface{}, error) {
 	cfg := types.Silenced{}
 	if err := unmarshalBody(req, &cfg); err != nil {
 		return nil, err
 	}
 
-	err := r.controller.Update(req.Context(), cfg)
+	err := r.controller.CreateOrReplace(req.Context(), cfg)
 	return cfg, err
 }