+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
 ### Added
 - Added ability to sort events in web UI.
+- Add PUT support to APId for the various resource types.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024